requirements.txt
python-dotenv
openai
requests
boto3
mistralai
